Move about dialog to core
