+++ packages/core/ui/AboutDialog.tsx
