+++ products/jbrowse-web/src/sessionModelFactory.ts
@@ -33,6 +33,7 @@ import RpcManager from '@jbrowse/core/rpc/RpcManager'
 import SettingsIcon from '@material-ui/icons/Settings'
 import CopyIcon from '@material-ui/icons/FileCopy'
 import DeleteIcon from '@material-ui/icons/Delete'
+import InfoIcon from '@material-ui/icons/Info'
 import shortid from 'shortid'
 
 declare interface ReferringNode {
@@ -88,6 +89,8 @@ export default function sessionModelFactory(
        * `{ taskName: "configure", target: thing_being_configured }`
        */
       task: undefined,
+
+      showAboutConfig: undefined as any,
     }))
     .views(self => ({
       get shareURL() {
@@ -175,6 +178,9 @@ export default function sessionModelFactory(
     .actions(self => ({
       setName(str: string) {
         self.name = str
+      },
+      setShowAboutConfig(showConfig: any) {
+        self.showAboutConfig = showConfig
       },
       addAssembly(assemblyConfig: any) {
         self.sessionAssemblies.push(assemblyConfig)
@@ -598,7 +604,7 @@ export default function sessionModelFactory(
         const session = self
         const canEdit =
           session.adminMode ||
-          session.sessionTracks.find((track: AnyConfigurationModel) => {
+          session.sessionTracks.find(track => {
             return track.trackId === config.trackId
           })
 
@@ -606,6 +612,13 @@ export default function sessionModelFactory(
         const isRefSeqTrack =
           readConfObject(config, 'type') === 'ReferenceSequenceTrack'
         return [
+          {
+            label: 'About track',
+            onClick: () => {
+              session.setShowAboutConfig(config)
+            },
+            icon: InfoIcon,
+          },
           {
             label: 'Settings',
             disabled: !canEdit,
